Support inheriting exported packages from regions defined earlier in the feature
When exported packages are defined in a region, these are inherited by
subsequent regions in the feature file. This is with the exception of
the global region, which is 'special'.
So if you have:
api-regions:JSON|false": [
  {"name": "global",
   "exports": ["com.global"]},
  {"name": "deprecated",
   "exports": ["some.package"]},
  {"name": "internal",
   "exports": ["dont.touch"]}]
Then the actual packages in the regions are as follows:
global: com.global
deprecated: some.package
internal: dont.touch, some.package