SLING-9521 Packages exported in earlier API Regions are not available to later API Regions
To support this the launcher extension now puts a property in the
features.properties map that contains the total ordering of regions,
e.g.:
  __region.order__=global,deprecated,internal
This commit also fixes the fact that the region values in the
features.properties map were unordered. Features have an order and this
is now preserved.